+++ .gitignore
@@ -1,6 +1,3 @@
 .npm
 node_modules
 .DS_Store
-directives
-browser.js
-index.js
