+++ src/Part.js
