+++ src/escape.js
