+++ src/is.js
