+++ src/promiseAsyncHtmlTemplate.js
