+++ src/streamAsyncHtmlTemplate.js
