+++ src/string.js
